Add Gate admin-level to tabs users and raports

+++ app/Http/Controllers/Admin/UserController.php
@@ -8,12 +8,15 @@ use App\Interfaces\UserRepositoryInterface;
 use App\Mail\TempPassChange;
 use App\Models\User;
 use App\Models\UserTypes;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Str;
 use Illuminate\View\View;
+use Throwable;
 
 class UserController extends Controller
 {
@@ -26,6 +29,8 @@ class UserController extends Controller
 
     public function index(Request $request): View
     {
+        Gate::authorize('admin-level');
+
         $phrase = $request->get('phrase');
         $email = $request->get('email');
         $phone = $request->get('phone');
@@ -58,6 +63,8 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('admin-level');
+
         $user = User::with('type')->findOrFail($id);
         $types = UserTypes::get();
         $commission_servies = $this->userRepository::COMMISSION_SERVIES;
@@ -73,6 +80,8 @@ class UserController extends Controller
 
     public function users_list(Request $request): View
     {
+        Gate::authorize('admin-level');
+
         $phrase = $request->get('phrase');
         $email = $request->get('email');
         $phone = $request->get('phone');
@@ -95,6 +104,8 @@ class UserController extends Controller
 
     public function create()
     {
+        Gate::authorize('admin-level');
+
         $types = UserTypes::get();
         $commission_servies = $this->userRepository::COMMISSION_SERVIES;
         $commission_medicals = $this->userRepository::COMMISSION_MEDICALS;
@@ -108,6 +119,8 @@ class UserController extends Controller
 
     public function store(AddUser $request)
     {
+        Gate::authorize('admin-level');
+
         $user = Auth::user();
 
         $data = $request->validated();
@@ -143,6 +156,8 @@ class UserController extends Controller
 
     public function update(AddUser $request, int $userId)
     {
+        Gate::authorize('admin-level');
+
         $data = $request->validated();
 
         $this->userRepository->update($data, $userId);
@@ -152,6 +167,8 @@ class UserController extends Controller
 
     public function changeStatus(int $id)
     {
+        Gate::authorize('admin-level');
+
         if ($id == Auth::id()) {
             return redirect()->back()->with('warning', 'Nie udało się, ponieważ swojego konta nie możesz wyłączyć!');
         }
@@ -162,6 +179,8 @@ class UserController extends Controller
 
     public function show(Request $request, int $userId): View
     {
+        Gate::authorize('admin-level');
+
         $loggedUser = Auth::user();
         $user = $this->userRepository->get($userId);
 
@@ -179,6 +198,8 @@ class UserController extends Controller
 
     public function delete($userId)
     {
+        Gate::authorize('admin-level');
+
         $this->userRepository->delete($userId);
 
         return redirect()->route('users.list')->with('success', 'Użytkownik został usunięty!');
